Completed useAuth hook :closed_lock_with_key:
